{{cookiecutter.project_name}}/requirements.txt
python-dateutil == 2.6.0
mongoengine == 0.11.0
flask == 1.0.2
flask_restful == 0.3.6
flask-script == 2.0.6
requests >= 2.13.0
Flask-JWT >= 0.3.2
openpyxl == 2.4.8
apscheduler == 3.3.1
flask_apidoc == 1.1.2
xpinyin == 0.5.5
nose == 1.3.7
mongomock == 3.10.0
nosexcover == 1.0.11
